+++ packages/passkey-controller/src/constants.ts
@@ -1,1 +1,21 @@
 export const PASSKEY_HKDF_INFO = 'MetaMask-passkey-encryption-v1';
+
+/**
+ * COSE Algorithms
+ *
+ * https://www.iana.org/assignments/cose/cose.xhtml#algorithms
+ */
+export enum COSEALG {
+  ES256 = -7,
+  EdDSA = -8,
+  ES384 = -35,
+  ES512 = -36,
+  PS256 = -37,
+  PS384 = -38,
+  PS512 = -39,
+  ES256K = -47,
+  RS256 = -257,
+  RS384 = -258,
+  RS512 = -259,
+  RS1 = -65535,
+}