+++ packages/passkey-controller/src/index.ts
@@ -6,6 +6,7 @@ export type {
   PasskeyControllerState,
   PasskeyControllerMessenger,
 } from './PasskeyController';
+export * from './constants';
 export * from './encoding';
 export type * from './types';
 export * from './webauthn';